# docs/openapi.yaml
openapi: 3.0.3
info:
  version : 1.0.0
  title: 콘서트 예약 서비스 API
  description: |
    콘서트 예약 서비스를 위한 RESTful API입니다.
...
paths:
  /api/v1/queue/token:
    post:
      tags:
        - Queue
      summary: 대기열 토큰 발급
      description: |
        사용자가 서비스를 이용하기 위한 대기열 토큰을 발급합니다.
        기존에 유효한 토큰이 있는 경우 해당 토큰을 반환합니다.
      operationId: issueQueueToken
      parameters:
        - name: userId
          in: query
          required: true
          description: 사용자 고유 식별자
          schema:
            type: string
            example: "HIDEONBUSH"
      responses:
        '200':
          description: 토큰 발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/queue/status:
    get:
      tags:
        - Queue
      summary: 대기열 상태 조회
      description: |
        사용자의 현재 대기열 상태를 조회합니다.
        대기 순서, 총 대기자 수, 남은 시간 등의 정보를 제공합니다.
      operationId: getQueueStatus
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-09876543"
      responses:
        '200':
          description: 대기열 상태 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStatusResponse'
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/concerts:
    get:
      tags:
        - Concert
      summary: 예약 가능한 콘서트 목록 조회
      description: |
        예약 가능한 콘서트 목록을 조회합니다.
        날짜 범위를 지정하여 특정 기간의 콘서트만 조회할 수 있습니다.
      operationId: getAvailableConcerts
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
        - name: startDate
          in: query
          required: false
          description: 조회 시작 날짜 (yyyy-MM-dd)
          schema:
            type: string
            format: date
            example: "2025-09-05"
        - name: endDate
          in: query
          required: false
          description: 조회 종료 날짜 (yyyy-MM-dd)
          schema:
            type: string
            format: date
            example: "2025-09-11"
      responses:
        '200':
          description: 콘서트 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  concerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConcertResponse'
                  totalCount:
                    type: integer
                    description: 총 콘서트 수
                    example: 5
              examples:
                multiple_concerts:
                  summary: 여러 콘서트 목록
                  value:
                    concerts:
                      - concertId: "RAIKASPACEDOG"
                        concertName: "우주개라이카"
                        concertDate: "2025-09-07"
                        totalSeatCount: 50
                        ticketPrice: 89000.00
                        isActive: true
                      - concertId: "WOMAN"
                        concertName: "미녀는괴로워"
                        concertDate: "2025-09-08"
                        totalSeatCount: 100
                        ticketPrice: 120000.00
                        isActive: true
                      - concertId: "HOME"
                        concertName: "집으로"
                        concertDate: "2025-09-08"
                        totalSeatCount: 200
                        ticketPrice: 75000.00
                        isActive: true
                      - concertId: "JOKER"
                        concertName: "조커"
                        concertDate: "2025-09-09"
                        totalSeatCount: 80
                        ticketPrice: 150000.00
                        isActive: true
                      - concertId: "ROCK2025"
                        concertName: "돌잔치치 페스티벌"
                        concertDate: "2025-09-11"
                        totalSeatCount: 500
                        ticketPrice: 200000.00
                        isActive: true
                    totalCount: 5
                single_concert:
                  summary: 단일 콘서트
                  value:
                    concerts:
                      - concertId: "RAIKASPACEDOG"
                        concertName: "우주개라이카"
                        concertDate: "2025-09-07"
                        totalSeatCount: 50
                        ticketPrice: 89000.00
                        isActive: true
                    totalCount: 1
                inactive_concert:
                  summary: 비활성 콘서트 포함
                  value:
                    concerts:
                      - concertId: "SUMMER2025"
                        concertName: "여름밤 콘서트"
                        concertDate: "2025-07-20"
                        totalSeatCount: 150
                        ticketPrice: 95000.00
                        isActive: false
                    totalCount: 1                    
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/concerts/dates:
    get:
      tags:
        - Concert
      summary: 예약 가능한 콘서트 날짜 목록 조회
      description: |
        예약 가능한 콘서트가 있는 날짜 목록을 조회합니다.
        날짜만 필요한 경우 사용하는 간소화된 API입니다. 개선되어 삭제될 예정입니다.
      operationId: getAvailableConcertDates
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
      responses:
        '200':
          description: 콘서트 날짜 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  dates:
                    type: array
                    items:
                      type: string
                      format: date
                      example: "2025-09-07"
                  totalCount:
                    type: integer
                    description: 총 날짜 수
                    example: 10
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/concerts/seats/details:
    get:
      tags:
        - Concert
      summary: 콘서트 좌석 상세 정보 조회
      description: |
        특정 콘서트의 좌석 상세 정보를 조회합니다.
        각 좌석의 상태(예약가능, 임시배정, 판매완료)와 전체 좌석 현황을 제공합니다.
      operationId: getConcertSeatDetails
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
        - name: concertDate
          in: query
          required: true
          description: 콘서트 날짜 (yyyy-MM-dd)
          schema:
            type: string
            format: date
            example: "2025-09-07"
        - name: concertTitle
          in: query
          required: true
          description: 콘서트 제목
          schema:
            type: string
            example: "우주개라이카"
      responses:
        '200':
          description: 좌석 상세 정보 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConcertSeatResponse'
        '400':
          description: 잘못된 요청 (존재하지 않는 콘서트)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/concerts/seats:
    get:
      tags:
        - Concert
      summary: 예약 가능한 좌석 목록 조회
      description: |
        특정 날짜의 예약 가능한 좌석 ID 목록을 조회합니다.
        간단한 좌석 목록만 필요한 경우 사용하는 API입니다.
      operationId: getAvailableSeats
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
        - name: concertDate
          in: query
          required: true
          description: 콘서트 날짜 (yyyy-MM-dd)
          schema:
            type: string
            format: date
            example: "2025-09-07"
      responses:
        '200':
          description: 예약 가능한 좌석 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  availableSeats:
                    type: array
                    items:
                      type: string
                      example: "1"
                  totalCount:
                    type: integer
                    description: 예약 가능한 좌석 수
                    example: 45
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/reservations:
    post:
      tags:
        - Reservation
      summary: 좌석 예약 요청
      description: |
        좌석을 예약하고 5분간 임시 배정합니다.
        임시 배정된 좌석은 다른 사용자가 예약할 수 없으며,
        5분 내에 결제를 완료해야 합니다.
      operationId: makeReservation
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '200':
          description: 예약 요청 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          description: 잘못된 요청 (입력값 검증 실패, 일반적인 예약 오류)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "RESERVATION_ERROR"
                message: "콘서트 날짜는 필수입니다."
                timestamp: "2025-09-05T10:00:00Z"
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "QUEUE_TOKEN_ERROR"
                message: "유효하지 않은 토큰입니다."
                timestamp: "2025-09-05T10:00:00Z"
        '410':
          description: 좌석이 존재하지 않거나 이미 판매됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "SEAT_NOT_FOUND"
                message: "존재하지 않는 좌석입니다."
                timestamp: "2025-09-05T10:00:00Z"
        '423':
          description: 좌석이 이미 임시 배정됨 (다른 사용자가 예약 중)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "SEAT_ALREADY_HELD"
                message: "이미 다른 사용자에게 임시배정된 좌석입니다."
                timestamp: "2025-09-05T10:00:00Z"
        '425':
          description: 대기열 비활성 상태 (아직 활성화되지 않음)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "NOT_ACTIVE_IN_QUEUE"
                message: "아직 활성화되지 않았습니다. 잠시만 기다려 주세요. (대기 인원 : 5)"
                timestamp: "2025-09-05T10:00:00Z"
        '498':
          description: 유효하지 않은 토큰 (커스텀 토큰 오류)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "INVALID_TOKEN"
                message: "유효하지 않은 토큰입니다."
                timestamp: "2025-09-05T10:00:00Z"
        '499':
          description: 토큰 만료 (커스텀 토큰 만료)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errorCode: "TOKEN_EXPIRED"
                message: "토큰이 만료되었습니다. 다시 발급받아 주세요."
                timestamp: "2025-09-05T10:00:00Z"
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - User
      summary: 사용자 예약 목록 조회
      description: |
        사용자의 모든 예약 목록을 조회합니다.
      operationId: getUserReservations
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
      responses:
        '200':
          description: 예약 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReservationResponse'
                  totalCount:
                    type: integer
                    description: 총 예약 수
                    example: 3
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/reservations/{reservationId}:
    get:
      tags:
        - Reservation
      summary: 예약 정보 조회
      description: |
        특정 예약의 상세 정보를 조회합니다.
        본인의 예약만 조회할 수 있습니다.
      operationId: getReservation
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
        - name: reservationId
          in: path
          required: true
          description: 예약 ID
          schema:
            type: string
            example: "RHIDE-12345"
      responses:
        '200':
          description: 예약 정보 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'

        '401':
          description: 유효하지 않은 토큰 또는 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/balance/charge:
    post:
      tags:
        - User
      summary: 잔액 충전
      description: |
        사용자의 잔액을 충전합니다.
        충전 내역이 거래 내역에 기록됩니다.
      operationId: chargeBalance
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BalanceChargeRequest'
      responses:
        '200':
          description: 잔액 충전 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceChargeResponse'

        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/balance:
    get:
      tags:
        - User
      summary: 잔액 조회
      description: |
        사용자의 현재 잔액과 최근 거래 내역을 조회합니다.
      operationId: getBalance
      parameters:
        - name: X-QUEUE-TOKEN
          in: header
          required: true
          description: 대기열 토큰 ID
          schema:
            type: string
            example: "THIDE-67890"
        - name: limit
          in: query
          required: false
          description: 조회할 거래 내역 수 
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
            example: 10
      responses:
        '200':
          description: 잔액 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          description: 유효하지 않은 토큰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    TokenResponse:
      type: object
      required:
        - tokenId
        - queueStatus
      properties:
        tokenId:
          type: string
          description: 발급된 토큰 ID
          example: "TOKEN+UUID12983789"
        queueStatus:
          $ref: '#/components/schemas/QueueStatusResponse'
    QueueStatusResponse:
      type: object
      required:
        - queuePosition
        - totalWaitingUsers
        - remainingTimeInSeconds
        - isActive
      properties:
        queuePosition:
          type: integer
          description: 현재 대기 순서
          example: 77
        totalWaitingUsers:
          type: integer
          description: 총 대기자 수
          example: 88
        remainingTimeInSeconds:
          type: integer
          description: 토큰 남은 유효 시간 (초)
          example: 299
        isActive:
          type: boolean
          description: 활성화 상태 여부
          example: false
    ConcertResponse:
      type: object
      required:
        - concertId
        - concertName
        - concertDate
        - totalSeatCount
        - ticketPrice
        - isActive
      properties:
        concertId:
          type: string
          description: 콘서트 ID
          example: "RAIKASPACEDOG"
        concertName:
          type: string
          description: 콘서트 이름
          example: "우주개라이카"
        concertDate:
          type: string
          format: date
          description: 콘서트 날짜
          example: "2025-09-07"
        totalSeatCount:
          type: integer
          description: 총 좌석 수
          example: 50
        ticketPrice:
          type: number
          format: decimal
          description: 티켓 가격
          example: 89000.00
        isActive:
          type: boolean
          description: 예약 가능 여부
          example: true
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: 오류 코드
          example: "INVALID_TOKEN"
        message:
          type: string
          description: 오류 메시지
          example: "유효하지 않은 토큰입니다."
        timestamp:
          type: string
          format: date-time
          description: 오류 발생 시간
          example: "2025-09-06T10:00:00"
        details:
          type: object
          description: 추가 오류 정보
          example:
            field: "tokenId"
            value: "invalid-token"
  securitySchemes:
    QueueTokenAuth:
      type: apiKey
      in: query
      name: tokenId
      description: 대기열 토큰을 통한 인증
...